moving UndoEngine.UndoUnit sub-classes to their own files
