+++ src/System.Windows.Forms.Design/src/System/ComponentModel/Design/UndoEngine.UndoUnit.cs
@@ -1,9 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
 
 using System.Collections;
-using System.ComponentModel.Design.Serialization;
 using System.Reflection;
 using System.Windows.Forms;
 
@@ -14,7 +12,7 @@ public abstract partial class UndoEngine
     /// <summary>
     ///  This class embodies a unit of undoable work.  The undo engine creates an undo unit when a change to the designer is about to be made.  The undo unit is responsible for tracking changes.  The undo engine will call Close on the unit when it no longer needs to track changes.
     /// </summary>
-    protected class UndoUnit
+    protected partial class UndoUnit
     {
         private List<UndoEvent>? _events; // the list of events we've captured
         private List<ChangeUndoEvent>? _changeEvents; // the list of change events we're currently capturing.  Only valid until Commit is called.
@@ -490,295 +488,6 @@ public abstract partial class UndoEngine
             }
 
             _reverse = !_reverse;
-        }
-
-        /// <summary>
-        ///  This undo event handles addition and removal of components.
-        /// </summary>
-        private sealed class AddRemoveUndoEvent : UndoEvent
-        {
-            private readonly SerializationStore _serializedData;
-            private readonly string? _componentName;
-
-            /// <summary>
-            ///  Creates a new object that contains the state of the event.  The last parameter, add, determines the initial mode of this event.  If true, it means this event is being created in response to a component add.  If false, it is being created in response to   a component remove.
-            /// </summary>
-            public AddRemoveUndoEvent(UndoEngine engine, IComponent component, bool add)
-            {
-                _componentName = component.Site!.Name;
-                NextUndoAdds = !add;
-                OpenComponent = component;
-
-                Debug.WriteLineIf(s_traceUndo.TraceVerbose, $"UndoEngine: ---> Creating {(add ? "Add" : "Remove")} undo event for '{_componentName}'");
-                using (_serializedData = engine._serializationService.CreateStore())
-                {
-                    engine._serializationService.Serialize(_serializedData, component);
-                }
-
-                // For add events, we commit as soon as we receive the event.
-                Committed = add;
-            }
-
-            /// <summary>
-            ///  Returns true if the add remove event has been comitted.
-            /// </summary>
-            internal bool Committed { get; private set; }
-
-            /// <summary>
-            ///  If this add/remove event is still open, OpenComponent will contain the component it is operating on.
-            /// </summary>
-            internal IComponent OpenComponent { get; }
-
-            /// <summary>
-            ///  Returns true if undoing this event will add a component.
-            /// </summary>
-            internal bool NextUndoAdds { get; private set; }
-
-            /// <summary>
-            ///  Commits this event.
-            /// </summary>
-            internal void Commit(UndoEngine engine)
-            {
-                if (!Committed)
-                {
-                    Debug.WriteLineIf(s_traceUndo.TraceVerbose, $"UndoEngine: ---> Committing remove of '{_componentName}'");
-                    Committed = true;
-                }
-            }
-
-            /// <summary>
-            ///  Actually performs the undo action.
-            /// </summary>
-            public override void Undo(UndoEngine engine)
-            {
-                if (NextUndoAdds)
-                {
-                    Debug.WriteLineIf(s_traceUndo.TraceVerbose, $"UndoEngine: ---> Adding '{_componentName}'");
-                    // We need to add this component.  To add it, we deserialize it and then we add it to the designer host's container.
-                    IDesignerHost host = engine.GetRequiredService<IDesignerHost>();
-
-                    engine._serializationService.DeserializeTo(_serializedData, host.Container);
-                }
-                else
-                {
-                    Debug.WriteLineIf(s_traceUndo.TraceVerbose, $"UndoEngine: ---> Removing '{_componentName}'");
-                    // We need to remove this component.  Take the name and match it to an object, and then ask that object to delete itself.
-                    IDesignerHost host = engine.GetRequiredService<IDesignerHost>();
-
-                    IComponent? component = host.Container.Components[_componentName];
-
-                    // Note: It's ok for the component to be null here.  This could happen if the parent to this control is disposed first. Ex:SplitContainer
-                    if (component is not null)
-                    {
-                        host.DestroyComponent(component);
-                    }
-                }
-
-                NextUndoAdds = !NextUndoAdds;
-            }
-        }
-
-        private sealed class ChangeUndoEvent : UndoEvent
-        {
-            // Static data we hang onto about this change.
-            private readonly string _componentName;
-            private readonly MemberDescriptor? _member;
-            // Before and after state.  Before state is built in the constructor. After state is built right before we undo for the first time.
-            private SerializationStore? _before;
-            private SerializationStore? _after;
-            private bool _savedAfterState;
-
-            /// <summary>
-            ///  Creates a new component change undo event.  This event consists of a before and after snapshot of a single component.  A snapshot will not be taken if a name for the component cannot be determined.
-            /// </summary>
-            public ChangeUndoEvent(UndoEngine engine, ComponentChangingEventArgs e, bool serializeBeforeState)
-            {
-                _componentName = engine.GetName(e.Component, true)!;
-                OpenComponent = e.Component;
-                _member = e.Member;
-
-                Debug.WriteLineIf(s_traceUndo.TraceVerbose, $"UndoEngine: ---> Creating change undo event for '{_componentName}'");
-                if (serializeBeforeState)
-                {
-                    Debug.WriteLineIf(s_traceUndo.TraceVerbose, $"UndoEngine: ---> Saving before snapshot for change to '{_componentName}'");
-                    _before = Serialize(engine, OpenComponent!, _member);
-                }
-            }
-
-            public ComponentChangingEventArgs ComponentChangingEventArgs => new(OpenComponent, _member);
-
-            /// <summary>
-            ///  Indicates that undoing this event may cause side effects in other objects.
-            ///  Change events fall into this category because, for example, a change involving adding an object to one collection may have a side effect of removing it from another collection.  Events with side effects are grouped at undo time so all their BeforeUndo methods are called before their Undo methods.
-            ///  Events without side effects have their BeforeUndo called and then their Undo called immediately after.
-            /// </summary>
-            public override bool CausesSideEffects => true;
-
-            /// <summary>
-            ///  Returns true if the change event has been comitted.
-            /// </summary>
-            [MemberNotNullWhen(false, nameof(OpenComponent))]
-            public bool Committed => OpenComponent is null;
-
-            /// <summary>
-            ///  Returns the component this change event is currently tracking. This will return null once the change event is committed.
-            /// </summary>
-            public object? OpenComponent { get; private set; }
-
-            /// <summary>
-            ///  Called before Undo is called. All undo events get their BeforeUndo called, and then they all get their Undo called. This allows the undo event to examine the state of the world before other undo events mess with it.
-            /// </summary>
-            public override void BeforeUndo(UndoEngine engine)
-            {
-                if (!_savedAfterState)
-                {
-                    _savedAfterState = true;
-                    SaveAfterState(engine);
-                }
-            }
-
-            /// <summary>
-            ///  Determines if this
-            /// </summary>
-            public bool ContainsChange(MemberDescriptor? desc)
-            {
-                if (_member is null)
-                {
-                    return true;
-                }
-
-                if (desc is null)
-                {
-                    return false;
-                }
-
-                return desc.Equals(_member);
-            }
-
-            /// <summary>
-            ///  Commits the unit.  Committing the unit saves the "after" snapshot of the unit.  If commit is called multiple times only the first commit is registered.
-            /// </summary>
-            public void Commit(UndoEngine engine)
-            {
-                if (!Committed)
-                {
-                    Debug.WriteLineIf(s_traceUndo.TraceVerbose, $"UndoEngine: ---> Committing change to '{_componentName}'");
-                    OpenComponent = null;
-                }
-            }
-
-            private void SaveAfterState(UndoEngine engine)
-            {
-                Debug.Assert(_after is null, "Change undo saving state twice.");
-                Debug.WriteLineIf(s_traceUndo.TraceVerbose, $"UndoEngine: ---> Saving after snapshot for change to '{_componentName}'");
-                object? component = null;
-
-                if (engine.TryGetService(out IReferenceService? rs))
-                {
-                    component = rs.GetReference(_componentName);
-                }
-                else if (engine.TryGetService(out IDesignerHost? host))
-                {
-                    component = host.Container.Components[_componentName];
-                }
-
-                // It is OK for us to not find a component here.  That can happen if our "after" state is owned by another change, like an add of the component.
-                if (component is not null)
-                {
-                    _after = Serialize(engine, component, _member);
-                }
-            }
-
-            private static SerializationStore Serialize(UndoEngine engine, object component, MemberDescriptor? member)
-            {
-                SerializationStore store;
-                using (store = engine._serializationService.CreateStore())
-                {
-                    if (member is not null && !(member.Attributes.Contains(DesignerSerializationVisibilityAttribute.Hidden)))
-                    {
-                        engine._serializationService.SerializeMemberAbsolute(store, component, member);
-                    }
-                    else
-                    {
-                        engine._serializationService.SerializeAbsolute(store, component);
-                    }
-                }
-
-                return store;
-            }
-
-            /// <summary>
-            ///  Performs the actual undo.  AFter it finishes it will reverse the role of _before and _after
-            /// </summary>
-            public override void Undo(UndoEngine engine)
-            {
-                Debug.WriteLineIf(s_traceUndo.TraceVerbose, $"UndoEngine: ---> Applying changes to '{_componentName}'");
-                Debug.Assert(_savedAfterState, "After state not saved.  BeforeUndo was not called?");
-
-                if (_before is not null)
-                {
-                    if (engine.TryGetService(out IDesignerHost? host))
-                    {
-                        engine._serializationService.DeserializeTo(_before, host.Container);
-                    }
-                }
-
-                (_after, _before) = (_before, _after);
-            }
-        }
-
-        private sealed class RenameUndoEvent : UndoEvent
-        {
-            private string? _before;
-            private string? _after;
-
-            /// <summary>
-            ///  Creates a new rename undo event.
-            /// </summary>
-            public RenameUndoEvent(string? before, string? after)
-            {
-                _before = before;
-                _after = after;
-                Debug.WriteLineIf(s_traceUndo.TraceVerbose, $"UndoEngine: ---> Creating rename undo event for '{_before}'->'{_after}'");
-            }
-
-            /// <summary>
-            ///  Simply undoes a rename by setting the name back to the saved value.
-            /// </summary>
-            public override void Undo(UndoEngine engine)
-            {
-                Debug.WriteLineIf(s_traceUndo.TraceVerbose, $"UndoEngine: ---> Renaming '{_after}'->'{_before}'");
-                IComponent? comp = engine._host.Container.Components[_after];
-                if (comp is not null)
-                {
-                    engine.ComponentChangeService.OnComponentChanging(comp, null);
-                    comp.Site!.Name = _before;
-                    (_after, _before) = (_before, _after);
-                }
-            }
-        }
-
-        private abstract class UndoEvent
-        {
-            /// <summary>
-            ///  Indicates that undoing this event may cause side effects in other objects.
-            ///  Change events fall into this category because, for example, a change involving adding an object to one collection may have a side effect of removing it from another collection.
-            ///  Events with side effects are grouped at undo time so all their BeforeUndo methods are called before their Undo methods.
-            ///  Events without side effects have their BeforeUndo called and then their Undo called immediately after.
-            /// </summary>
-            public virtual bool CausesSideEffects => false;
-
-            /// <summary>
-            ///  Called before Undo is called.  All undo events get their BeforeUndo called, and then they all get their Undo called. This allows the undo event to examine the state of the world before other undo events mess with it. BeforeUndo returns true if before undo was supported, and false if not.  If before undo is not supported, the undo unit should be undone immediately.
-            /// </summary>
-            public virtual void BeforeUndo(UndoEngine engine)
-            {
-            }
-
-            /// <summary>
-            ///  Called by the undo unit when it wants to undo this bit of work.
-            /// </summary>
-            public abstract void Undo(UndoEngine engine);
         }
     }
 }
